Separated TAS_Pizza_Factions config.cpp into multiple files
