+++ addons/TAS_Pizza_Factions/CfgGroups.hpp
@@ -1,0 +1,804 @@
+class CfgGroups
+{
+	//********************************************************************************************************************************************************************************************
+    //*****            Dominos Groups              *****************************************************************************************************************************************************
+    //********************************************************************************************************************************************************************************************
+	class West
+	{
+		class TAS_Dominos_Zimbabwe
+		{
+			name = "Dominos Zimbabwe";
+			//********************************************************************************************************************************************************************************************
+			//*****            Dominos Infantry Groups              *****************************************************************************************************************************************************
+			//********************************************************************************************************************************************************************************************
+			class Infantry
+			{
+				name = "Valued Employees";
+				class TAS_Dominos_EmployeeTeam
+				{
+					name = "Employee Team";
+					side = 1;
+					faction = "TAS_Dominos_Zimbabwe";
+					icon = "\A3\ui_f\data\map\markers\nato\b_inf.paa";
+					class Unit0
+					{
+						side = 1;
+						vehicle = "TAS_Dominos_Employee";
+						rank = "SERGEANT";
+						position[] = {0, 0, 0};
+					};
+					class Unit1
+					{
+						side = 1;
+						vehicle = "TAS_Dominos_Employee";
+						rank = "PRIVATE";
+						position[] = {5, -5, 0};
+					};
+					class Unit2
+					{
+						side = 1;
+						vehicle = "TAS_Dominos_Employee";
+						rank = "PRIVATE";
+						position[] = {-5, -5, 0};
+					};
+					class Unit3
+					{
+						side = 1;
+						vehicle = "TAS_Dominos_Employee";
+						rank = "PRIVATE";
+						position[] = {10, -10, 0};
+					};
+					class Unit4
+					{
+						side = 1;
+						vehicle = "TAS_Dominos_Employee";
+						rank = "PRIVATE";
+						position[] = {-10, -10, 0};
+					};
+				};
+				class TAS_Dominos_EmployeeCombatTeam
+				{
+					name = "Employee Combat Team";
+					side = 1;
+					faction = "TAS_Dominos_Zimbabwe";
+					icon = "\A3\ui_f\data\map\markers\nato\b_inf.paa";
+					class Unit0
+					{
+						side = 1;
+						vehicle = "TAS_Dominos_EmployeeCombat";
+						rank = "SERGEANT";
+						position[] = {0, 0, 0};
+					};
+					class Unit1
+					{
+						side = 1;
+						vehicle = "TAS_Dominos_EmployeeCombat";
+						rank = "PRIVATE";
+						position[] = {5, -5, 0};
+					};
+					class Unit2
+					{
+						side = 1;
+						vehicle = "TAS_Dominos_EmployeeCombat";
+						rank = "PRIVATE";
+						position[] = {-5, -5, 0};
+					};
+					class Unit3
+					{
+						side = 1;
+						vehicle = "TAS_Dominos_EmployeeCombat";
+						rank = "PRIVATE";
+						position[] = {10, -10, 0};
+					};
+					class Unit4
+					{
+						side = 1;
+						vehicle = "TAS_Dominos_EmployeeCombat";
+						rank = "PRIVATE";
+						position[] = {-10, -10, 0};
+					};
+				};
+				class TAS_Dominos_ManagerTeam
+				{
+					name = "Manager Team";
+					side = 1;
+					faction = "TAS_Dominos_Zimbabwe";
+					icon = "\A3\ui_f\data\map\markers\nato\b_inf.paa";
+					class Unit0
+					{
+						side = 1;
+						vehicle = "TAS_Dominos_Manager";
+						rank = "SERGEANT";
+						position[] = {0, 0, 0};
+					};
+					class Unit1
+					{
+						side = 1;
+						vehicle = "TAS_Dominos_Manager";
+						rank = "PRIVATE";
+						position[] = {5, -5, 0};
+					};
+					class Unit2
+					{
+						side = 1;
+						vehicle = "TAS_Dominos_Manager";
+						rank = "PRIVATE";
+						position[] = {-5, -5, 0};
+					};
+					class Unit3
+					{
+						side = 1;
+						vehicle = "TAS_Dominos_Manager";
+						rank = "PRIVATE";
+						position[] = {10, -10, 0};
+					};
+					class Unit4
+					{
+						side = 1;
+						vehicle = "TAS_Dominos_Manager";
+						rank = "PRIVATE";
+						position[] = {-10, -10, 0};
+					};
+				};
+				class TAS_Dominos_OperativeTeam
+				{
+					name = "Operative Team";
+					side = 1;
+					faction = "TAS_Dominos_Zimbabwe";
+					icon = "\A3\ui_f\data\map\markers\nato\b_inf.paa";
+					class Unit0
+					{
+						side = 1;
+						vehicle = "TAS_Dominos_Operative";
+						rank = "SERGEANT";
+						position[] = {0, 0, 0};
+					};
+					class Unit1
+					{
+						side = 1;
+						vehicle = "TAS_Dominos_Operative";
+						rank = "PRIVATE";
+						position[] = {5, -5, 0};
+					};
+					class Unit2
+					{
+						side = 1;
+						vehicle = "TAS_Dominos_Operative";
+						rank = "PRIVATE";
+						position[] = {-5, -5, 0};
+					};
+					class Unit3
+					{
+						side = 1;
+						vehicle = "TAS_Dominos_Operative";
+						rank = "PRIVATE";
+						position[] = {10, -10, 0};
+					};
+					class Unit4
+					{
+						side = 1;
+						vehicle = "TAS_Dominos_Operative";
+						rank = "PRIVATE";
+						position[] = {-10, -10, 0};
+					};
+				};
+				class TAS_Dominos_ShiftTeam
+				{
+					name = "Minimum Wage Team";
+					side = 1;
+					faction = "TAS_Dominos_Zimbabwe";
+					icon = "\A3\ui_f\data\map\markers\nato\b_inf.paa";
+					class Unit0
+					{
+						side = 1;
+						vehicle = "TAS_Dominos_Manager";
+						rank = "SERGEANT";
+						position[] = {0, 0, 0};
+					};
+					class Unit1
+					{
+						side = 1;
+						vehicle = "TAS_Dominos_Employee";
+						rank = "PRIVATE";
+						position[] = {5, -5, 0};
+					};
+					class Unit2
+					{
+						side = 1;
+						vehicle = "TAS_Dominos_Employee";
+						rank = "PRIVATE";
+						position[] = {-5, -5, 0};
+					};
+					class Unit3
+					{
+						side = 1;
+						vehicle = "TAS_Dominos_Employee";
+						rank = "PRIVATE";
+						position[] = {10, -10, 0};
+					};
+					class Unit4
+					{
+						side = 1;
+						vehicle = "TAS_Dominos_Employee";
+						rank = "PRIVATE";
+						position[] = {-10, -10, 0};
+					};
+				};
+				class TAS_Dominos_ShiftCombatTeam
+				{
+					name = "Hazard Pay Team";
+					side = 1;
+					faction = "TAS_Dominos_Zimbabwe";
+					icon = "\A3\ui_f\data\map\markers\nato\b_inf.paa";
+					class Unit0
+					{
+						side = 1;
+						vehicle = "TAS_Dominos_Manager";
+						rank = "SERGEANT";
+						position[] = {0, 0, 0};
+					};
+					class Unit1
+					{
+						side = 1;
+						vehicle = "TAS_Dominos_EmployeeCombat";
+						rank = "PRIVATE";
+						position[] = {5, -5, 0};
+					};
+					class Unit2
+					{
+						side = 1;
+						vehicle = "TAS_Dominos_EmployeeCombat";
+						rank = "PRIVATE";
+						position[] = {-5, -5, 0};
+					};
+					class Unit3
+					{
+						side = 1;
+						vehicle = "TAS_Dominos_EmployeeCombat";
+						rank = "PRIVATE";
+						position[] = {10, -10, 0};
+					};
+					class Unit4
+					{
+						side = 1;
+						vehicle = "TAS_Dominos_EmployeeCombat";
+						rank = "PRIVATE";
+						position[] = {-10, -10, 0};
+					};
+				};
+			};
+			//********************************************************************************************************************************************************************************************
+			//*****            Dominos Vehicle Groups              *****************************************************************************************************************************************************
+			//********************************************************************************************************************************************************************************************
+			class Vehicles
+			{
+				name = "Pizza Delivery Teams";
+				class TAS_Dominos_MotorizedOffroad
+				{
+					name = "Pizza Delivery Team";
+					side = 1;
+					faction = "TAS_Dominos_Zimbabwe";
+					icon = "\A3\ui_f\data\map\markers\nato\b_motor_inf.paa";
+					class Unit0
+					{
+						side = 1;
+						vehicle = "TAS_Dominos_Offroad";
+						rank = "SERGEANT";
+						position[] = {0, 0, 0};
+					};
+					class Unit1
+					{
+						side = 1;
+						vehicle = "TAS_Dominos_Manager";
+						rank = "CORPORAL";
+						position[] = {5, -5, 0};
+					};
+					class Unit2
+					{
+						side = 1;
+						vehicle = "TAS_Dominos_Employee";
+						rank = "PRIVATE";
+						position[] = {-5, -5, 0};
+					};
+					class Unit3
+					{
+						side = 1;
+						vehicle = "TAS_Dominos_Employee";
+						rank = "PRIVATE";
+						position[] = {10, -10, 0};
+					};
+					class Unit4
+					{
+						side = 1;
+						vehicle = "TAS_Dominos_Employee";
+						rank = "PRIVATE";
+						position[] = {-10, -10, 0};
+					};
+					class Unit5
+					{
+						side = 1;
+						vehicle = "TAS_Dominos_Employee";
+						rank = "PRIVATE";
+						position[] = {15, -15, 0};
+					};
+				};
+				class TAS_Dominos_MotorizedCombatOffroad
+				{
+					name = "Combat Pizza Delivery Team";
+					side = 1;
+					faction = "TAS_Dominos_Zimbabwe";
+					icon = "\A3\ui_f\data\map\markers\nato\b_motor_inf.paa";
+					class Unit0
+					{
+						side = 1;
+						vehicle = "TAS_Dominos_Offroad";
+						rank = "SERGEANT";
+						position[] = {0, 0, 0};
+					};
+					class Unit1
+					{
+						side = 1;
+						vehicle = "TAS_Dominos_Manager";
+						rank = "CORPORAL";
+						position[] = {5, -5, 0};
+					};
+					class Unit2
+					{
+						side = 1;
+						vehicle = "TAS_Dominos_EmployeeCombat";
+						rank = "PRIVATE";
+						position[] = {-5, -5, 0};
+					};
+					class Unit3
+					{
+						side = 1;
+						vehicle = "TAS_Dominos_EmployeeCombat";
+						rank = "PRIVATE";
+						position[] = {10, -10, 0};
+					};
+					class Unit4
+					{
+						side = 1;
+						vehicle = "TAS_Dominos_EmployeeCombat";
+						rank = "PRIVATE";
+						position[] = {-10, -10, 0};
+					};
+					class Unit5
+					{
+						side = 1;
+						vehicle = "TAS_Dominos_EmployeeCombat";
+						rank = "PRIVATE";
+						position[] = {15, -15, 0};
+					};
+				};
+				class TAS_Dominos_Technicals
+				{
+					name = "Pizza Delivery Technicals";
+					side = 1;
+					faction = "TAS_Dominos_Zimbabwe";
+					icon = "\A3\ui_f\data\map\markers\nato\b_motor_inf.paa";
+					class Unit0
+					{
+						side = 1;
+						vehicle = "TAS_Dominos_Offroad_HMG";
+						rank = "LIEUTENANT";
+						position[] = {0, 0, 0};
+					};
+					class Unit1
+					{
+						side = 1;
+						vehicle = "TAS_Dominos_Offroad_HMG";
+						rank = "SERGEANT";
+						position[] = {10, -10, 0};
+					};
+					class Unit2
+					{
+						side = 1;
+						vehicle = "TAS_Dominos_Offroad_AT";
+						rank = "CORPORAL";
+						position[] = {-10, -10, 0};
+					};
+				};
+			};
+		};		
+	};
+	class East
+	{
+		class TAS_Papa_Jhons
+		{
+			name = "Papa Jhons";
+			//********************************************************************************************************************************************************************************************
+			//*****            Papa Jhons Infantry Groups              *****************************************************************************************************************************************************
+			//********************************************************************************************************************************************************************************************
+			class Infantry
+			{
+				name = "Valued Employees";
+				class TAS_PapaJhons_EmployeeTeam
+				{
+					name = "Employee Team";
+					side = 0;
+					faction = "TAS_Papa_Jhons";
+					icon = "\A3\ui_f\data\map\markers\nato\b_inf.paa";
+					class Unit0
+					{
+						side = 0;
+						vehicle = "TAS_PapaJhons_Employee";
+						rank = "SERGEANT";
+						position[] = {0, 0, 0};
+					};
+					class Unit1
+					{
+						side = 0;
+						vehicle = "TAS_PapaJhons_Employee";
+						rank = "PRIVATE";
+						position[] = {5, -5, 0};
+					};
+					class Unit2
+					{
+						side = 0;
+						vehicle = "TAS_PapaJhons_Employee";
+						rank = "PRIVATE";
+						position[] = {-5, -5, 0};
+					};
+					class Unit3
+					{
+						side = 0;
+						vehicle = "TAS_PapaJhons_Employee";
+						rank = "PRIVATE";
+						position[] = {10, -10, 0};
+					};
+					class Unit4
+					{
+						side = 0;
+						vehicle = "TAS_PapaJhons_Employee";
+						rank = "PRIVATE";
+						position[] = {-10, -10, 0};
+					};
+				};
+				class TAS_PapaJhons_EmployeeCombatTeam
+				{
+					name = "Employee Combat Team";
+					side = 0;
+					faction = "TAS_Papa_Jhons";
+					icon = "\A3\ui_f\data\map\markers\nato\b_inf.paa";
+					class Unit0
+					{
+						side = 0;
+						vehicle = "TAS_PapaJhons_EmployeeCombat";
+						rank = "SERGEANT";
+						position[] = {0, 0, 0};
+					};
+					class Unit1
+					{
+						side = 0;
+						vehicle = "TAS_PapaJhons_EmployeeCombat";
+						rank = "PRIVATE";
+						position[] = {5, -5, 0};
+					};
+					class Unit2
+					{
+						side = 0;
+						vehicle = "TAS_PapaJhons_EmployeeCombat";
+						rank = "PRIVATE";
+						position[] = {-5, -5, 0};
+					};
+					class Unit3
+					{
+						side = 0;
+						vehicle = "TAS_PapaJhons_EmployeeCombat";
+						rank = "PRIVATE";
+						position[] = {10, -10, 0};
+					};
+					class Unit4
+					{
+						side = 0;
+						vehicle = "TAS_PapaJhons_EmployeeCombat";
+						rank = "PRIVATE";
+						position[] = {-10, -10, 0};
+					};
+				};
+				class TAS_PapaJhons_ManagerTeam
+				{
+					name = "Manager Team";
+					side = 0;
+					faction = "TAS_Papa_Jhons";
+					icon = "\A3\ui_f\data\map\markers\nato\b_inf.paa";
+					class Unit0
+					{
+						side = 0;
+						vehicle = "TAS_PapaJhons_Manager";
+						rank = "SERGEANT";
+						position[] = {0, 0, 0};
+					};
+					class Unit1
+					{
+						side = 0;
+						vehicle = "TAS_PapaJhons_Manager";
+						rank = "PRIVATE";
+						position[] = {5, -5, 0};
+					};
+					class Unit2
+					{
+						side = 0;
+						vehicle = "TAS_PapaJhons_Manager";
+						rank = "PRIVATE";
+						position[] = {-5, -5, 0};
+					};
+					class Unit3
+					{
+						side = 0;
+						vehicle = "TAS_PapaJhons_Manager";
+						rank = "PRIVATE";
+						position[] = {10, -10, 0};
+					};
+					class Unit4
+					{
+						side = 0;
+						vehicle = "TAS_PapaJhons_Manager";
+						rank = "PRIVATE";
+						position[] = {-10, -10, 0};
+					};
+				};
+				class TAS_PapaJhons_OperativeTeam
+				{
+					name = "Operative Team";
+					side = 0;
+					faction = "TAS_Papa_Jhons";
+					icon = "\A3\ui_f\data\map\markers\nato\b_inf.paa";
+					class Unit0
+					{
+						side = 0;
+						vehicle = "TAS_PapaJhons_Operative";
+						rank = "SERGEANT";
+						position[] = {0, 0, 0};
+					};
+					class Unit1
+					{
+						side = 0;
+						vehicle = "TAS_PapaJhons_Operative";
+						rank = "PRIVATE";
+						position[] = {5, -5, 0};
+					};
+					class Unit2
+					{
+						side = 0;
+						vehicle = "TAS_PapaJhons_Operative";
+						rank = "PRIVATE";
+						position[] = {-5, -5, 0};
+					};
+					class Unit3
+					{
+						side = 0;
+						vehicle = "TAS_PapaJhons_Operative";
+						rank = "PRIVATE";
+						position[] = {10, -10, 0};
+					};
+					class Unit4
+					{
+						side = 0;
+						vehicle = "TAS_PapaJhons_Operative";
+						rank = "PRIVATE";
+						position[] = {-10, -10, 0};
+					};
+				};
+				class TAS_PapaJhons_ShiftTeam
+				{
+					name = "Minimum Wage Team";
+					side = 0;
+					faction = "TAS_Papa_Jhons";
+					icon = "\A3\ui_f\data\map\markers\nato\b_inf.paa";
+					class Unit0
+					{
+						side = 0;
+						vehicle = "TAS_PapaJhons_Manager";
+						rank = "SERGEANT";
+						position[] = {0, 0, 0};
+					};
+					class Unit1
+					{
+						side = 0;
+						vehicle = "TAS_PapaJhons_Employee";
+						rank = "PRIVATE";
+						position[] = {5, -5, 0};
+					};
+					class Unit2
+					{
+						side = 0;
+						vehicle = "TAS_PapaJhons_Employee";
+						rank = "PRIVATE";
+						position[] = {-5, -5, 0};
+					};
+					class Unit3
+					{
+						side = 0;
+						vehicle = "TAS_PapaJhons_Employee";
+						rank = "PRIVATE";
+						position[] = {10, -10, 0};
+					};
+					class Unit4
+					{
+						side = 0;
+						vehicle = "TAS_PapaJhons_Employee";
+						rank = "PRIVATE";
+						position[] = {-10, -10, 0};
+					};
+				};
+				class TAS_PapaJhons_ShiftCombatTeam
+				{
+					name = "Hazard Pay Team";
+					side = 0;
+					faction = "TAS_Papa_Jhons";
+					icon = "\A3\ui_f\data\map\markers\nato\b_inf.paa";
+					class Unit0
+					{
+						side = 0;
+						vehicle = "TAS_PapaJhons_Manager";
+						rank = "SERGEANT";
+						position[] = {0, 0, 0};
+					};
+					class Unit1
+					{
+						side = 0;
+						vehicle = "TAS_PapaJhons_EmployeeCombat";
+						rank = "PRIVATE";
+						position[] = {5, -5, 0};
+					};
+					class Unit2
+					{
+						side = 0;
+						vehicle = "TAS_PapaJhons_EmployeeCombat";
+						rank = "PRIVATE";
+						position[] = {-5, -5, 0};
+					};
+					class Unit3
+					{
+						side = 0;
+						vehicle = "TAS_PapaJhons_EmployeeCombat";
+						rank = "PRIVATE";
+						position[] = {10, -10, 0};
+					};
+					class Unit4
+					{
+						side = 0;
+						vehicle = "TAS_PapaJhons_EmployeeCombat";
+						rank = "PRIVATE";
+						position[] = {-10, -10, 0};
+					};
+				};
+			};
+			//********************************************************************************************************************************************************************************************
+			//*****            Papa Jhons Vehicle Groups              *****************************************************************************************************************************************************
+			//********************************************************************************************************************************************************************************************
+			class Vehicles
+			{
+				name = "Pizza Delivery Teams";
+				class TAS_PapaJhons_MotorizedOffroad
+				{
+					name = "Pizza Delivery Team";
+					side = 0;
+					faction = "TAS_Papa_Jhons";
+					icon = "\A3\ui_f\data\map\markers\nato\b_motor_inf.paa";
+					class Unit0
+					{
+						side = 0;
+						vehicle = "TAS_PapaJhons_Offroad";
+						rank = "SERGEANT";
+						position[] = {0, 0, 0};
+					};
+					class Unit1
+					{
+						side = 0;
+						vehicle = "TAS_PapaJhons_Manager";
+						rank = "CORPORAL";
+						position[] = {5, -5, 0};
+					};
+					class Unit2
+					{
+						side = 0;
+						vehicle = "TAS_PapaJhons_Employee";
+						rank = "PRIVATE";
+						position[] = {-5, -5, 0};
+					};
+					class Unit3
+					{
+						side = 0;
+						vehicle = "TAS_PapaJhons_Employee";
+						rank = "PRIVATE";
+						position[] = {10, -10, 0};
+					};
+					class Unit4
+					{
+						side = 0;
+						vehicle = "TAS_PapaJhons_Employee";
+						rank = "PRIVATE";
+						position[] = {-10, -10, 0};
+					};
+					class Unit5
+					{
+						side = 0;
+						vehicle = "TAS_PapaJhons_Employee";
+						rank = "PRIVATE";
+						position[] = {15, -15, 0};
+					};
+				};
+				class TAS_PapaJhons_MotorizedCombatOffroad
+				{
+					name = "Combat Pizza Delivery Team";
+					side = 0;
+					faction = "TAS_Papa_Jhons";
+					icon = "\A3\ui_f\data\map\markers\nato\b_motor_inf.paa";
+					class Unit0
+					{
+						side = 0;
+						vehicle = "TAS_PapaJhons_Offroad";
+						rank = "SERGEANT";
+						position[] = {0, 0, 0};
+					};
+					class Unit1
+					{
+						side = 0;
+						vehicle = "TAS_PapaJhons_Manager";
+						rank = "CORPORAL";
+						position[] = {5, -5, 0};
+					};
+					class Unit2
+					{
+						side = 0;
+						vehicle = "TAS_PapaJhons_EmployeeCombat";
+						rank = "PRIVATE";
+						position[] = {-5, -5, 0};
+					};
+					class Unit3
+					{
+						side = 0;
+						vehicle = "TAS_PapaJhons_EmployeeCombat";
+						rank = "PRIVATE";
+						position[] = {10, -10, 0};
+					};
+					class Unit4
+					{
+						side = 0;
+						vehicle = "TAS_PapaJhons_EmployeeCombat";
+						rank = "PRIVATE";
+						position[] = {-10, -10, 0};
+					};
+					class Unit5
+					{
+						side = 0;
+						vehicle = "TAS_PapaJhons_EmployeeCombat";
+						rank = "PRIVATE";
+						position[] = {15, -15, 0};
+					};
+				};
+				class TAS_PapaJhons_Technicals
+				{
+					name = "Pizza Delivery Technicals";
+					side = 0;
+					faction = "TAS_Papa_Jhons";
+					icon = "\A3\ui_f\data\map\markers\nato\b_motor_inf.paa";
+					class Unit0
+					{
+						side = 0;
+						vehicle = "TAS_PapaJhons_Offroad_HMG";
+						rank = "LIEUTENANT";
+						position[] = {0, 0, 0};
+					};
+					class Unit1
+					{
+						side = 0;
+						vehicle = "TAS_PapaJhons_Offroad_HMG";
+						rank = "SERGEANT";
+						position[] = {10, -10, 0};
+					};
+					class Unit2
+					{
+						side = 0;
+						vehicle = "TAS_PapaJhons_Offroad_AT";
+						rank = "CORPORAL";
+						position[] = {-10, -10, 0};
+					};
+				};
+			};
+		};		
+	};
+};
